Initial fix for browsing URLs
This commit enhances `Browser` logic within `pkg/browser` to detect supported and possibly supported URL schemes / protocols.

If a URL is presented that doesn't match a known supported scheme/protocol, then `Browse(string) error` will check if the provided URL matches a file or directory on the filesystem.  Erring if this matches anything we can find on the filesystem.